OpenCLIP/visual-search/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==9.5.0
open-clip-torch==2.32.0
torch==2.1.0
pymilvus==2.3.3
mysql-connector-python==8.2.0
numpy==1.24.3
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
redis==5.0.1
slowapi==0.1.9
marshmallow==3.20.2
environs==9.5.0
requests==2.31.0
minio==7.2.0
jieba==0.42.1
tenacity==8.2.3
